Merged (manually) changes made on trunk commits 13504 and 13511
Crop sequence numbers were not matching definition of a crop interval. Moved triggers to include crop termination without a harvest. Now crop termination ends the crop interval. This makes a planted cover crop have it's own interval.

Further changes to make crop sequence numbers match the  definition of a crop interval. Tested to confirm works with termination by planting operation and termination of cover crop as first operation of a rotation. Note: may not work with -I1 and definitely not with -I0.

git-svn-id: https://infosys.ars.usda.gov/svn/code/weps1/branches/weps.src.subregion@13622 77af8a48-1994-4e4a-b6f4-b1c89642997b