Merged (manually) changes made on trunk checkins 13869 and 13898
Corrected problem with planting date back calculation for season.out where the leap days that occurred while the crop was growing were no the same as suring the period of the back calculation. Leap days are now tracked during crop growth and also tracked over the back calculation date range and corrected accordingly. The problem occured when a crop was grow for more than 1 year and the error appeared only on the first planting date backcalcuation.

Made check for crop type to be less than or equal to zero when setting "no growth state" for invalid crop.

git-svn-id: https://infosys.ars.usda.gov/svn/code/weps1/branches/weps.src.subregion@13902 77af8a48-1994-4e4a-b6f4-b1c89642997b